# docs/swagger/swagger.yaml
definitions:
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  handler.Message:
    type: object
  handler.MessageResponse:
    properties:
      data: {}
      status:
        type: string
    type: object
  handler.UserResponse:
    properties:
      sessionId:
        type: string
      subSessionId:
        type: string
      token:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/init:
    post:
      consumes:
      - application/json
      description: Creates a new user session and returns authentication tokens
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UserResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Initialize user session
      tags:
      - user
  /api/messages:
    post:
      consumes:
      - application/json
      description: Sends a new chat message
      parameters:
      - description: Message to send
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/handler.Message'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Send chat message
      tags:
      - messages
  /api/messages/offline:
    get:
      consumes:
      - application/json
      description: Retrieves all offline messages for the specified user
      parameters:
      - description: User ID
        in: query
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get offline messages
      tags:
      - messages
swagger: "2.0"